incremental checkin: refactoring file access methods out of observation uploader, preparing for ApiManager refactor. refactoring connection check (basely used). Javadoc'ing Api Manager
